backend/app/utils/security.py
from datetime import datetime, timedelta
from typing import Any, Optional

from jose import jwt
from passlib.context import CryptContext

from ..core.config import settings

pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")


# ───────── password hashing ─────────────────────────────────
def verify_password(plain: str, hashed: str) -> bool:
    return pwd_context.verify(plain, hashed)


def get_password_hash(password: str) -> str:
    return pwd_context.hash(password)


# ───────── JWT tokens ───────────────────────────────────────
def create_access_token(
    subject: str | dict[str, Any],
    expires_minutes: int = settings.ACCESS_TOKEN_EXPIRE_MINUTES,
) -> str:
    if isinstance(subject, str):
        data = {"sub": subject}
    else:
        data = subject
    expire = datetime.utcnow() + timedelta(minutes=expires_minutes)
    payload = {**data, "exp": expire}
    return jwt.encode(payload, settings.SECRET_KEY, algorithm=settings.ALGORITHM)
