backend/app/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.31
asyncpg==0.29.0          # Postgres driver
python-dotenv==1.0.1
pydantic==2.7.1
python-jose[cryptography]==3.3.0  # JWT
passlib[bcrypt]==1.7.4
stripe==9.5.0
python-multipart==0.0.9
aiofiles==23.2.1
httpx==0.27.0
